\name{gesel-package}
\alias{gesel-package}
\alias{gesel}
\docType{package}
\title{
\packageTitle{gesel}
}
\description{
\packageDescription{gesel}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{gesel}
\packageIndices{gesel}

}
\author{
\packageAuthor{gesel}

Maintainer: \packageMaintainer{gesel}
}